% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{release}
\alias{release}
\title{Collect local IASI release artifacts}
\usage{
release(path = ".")
}
\arguments{
\item{path}{Directory from which IASI projects are discovered.}
}
\value{
Numeric status bitmask.
}
\description{
Collect local IASI release artifacts
}
